Move the fucking check in the base class.
